\name{AF_XH-package}
\alias{AF_XH-package}
\alias{AF_XH}
\docType{package}
\title{
This is our package for Assignment 5.
}
\description{
In this package, an analysis of email data from several sources is implemented.
}
\details{
\tabular{ll}{
Package: \tab AF_XH\cr
Type: \tab Package\cr
Version: \tab 0.0-0\cr
Date: \tab 2011-02-26\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andrew Farris, Xinyi Hu, Yi Chen
}
\examples{
test1()
}
